Enviando arquivos faltante.
